fix: picks now block board load
