+++ app/leagues/[id]/draft/[draftId]/page.tsx
@@ -19,6 +19,86 @@ import { createClient } from '@/app/utils/supabase/server';
 import Board from '@/components/ui/draft/board';
 import { BoardProps, DraftPick } from '@/lib/types';
 import { User } from '@supabase/supabase-js';
+import { revalidatePath } from 'next/cache';
+import { cache } from 'react';
+
+type DraftProps = {
+   league: League;
+   draft: Draft;
+   watchlist: Watchlist;
+   user: User;
+   draftPicks: DraftPick[];
+   players: Player[];
+   team: Team;
+   teams: Team[];
+   leagueRules: LeagueRules;
+   leagueScoring: LeagueScoring;
+   draftedPlayers: DraftSelection[];
+   timerDuration: number;
+};
+
+const getDraftProps = cache(
+   async ({
+      id,
+      draftId,
+   }: {
+      id: string;
+      draftId: string;
+   }): Promise<DraftProps> => {
+      const supabase = createClient();
+
+      const user: Awaited<User | null> = await getUser(supabase);
+
+      if (!user) throw new Error('User is required');
+
+      const draft: Awaited<Draft> = await fetchDraftById(supabase, draftId);
+      const draftedPlayers: Awaited<DraftSelection[]> =
+         await fetchDraftedPlayers(supabase, draft);
+      const draftPicks: Awaited<DraftPick[]> = await fetchDraftPicks(
+         supabase,
+         draftId
+      );
+      const league: Awaited<League> = await fetchLeague(supabase, id);
+      const players: Awaited<Player[]> = await getPlayers(id);
+      const team: Awaited<Team> = await fetchTeam(supabase, user.id, id);
+      const leagueRules: Awaited<LeagueRules> = await fetchLeagueRules(
+         supabase,
+         league
+      );
+      const leagueScoring: Awaited<LeagueScoring> = await fetchLeagueScoring(
+         supabase,
+         league
+      );
+      const teams: Awaited<Team[]> = await fetchTeams(
+         supabase,
+         league.league_id as string
+      );
+      const timerDuration: Awaited<number> = await getTimerDuration(
+         supabase,
+         leagueRules.id
+      );
+      const watchlist: Awaited<Watchlist> = await fetchWatchlist(
+         supabase,
+         user.id,
+         draft
+      );
+
+      return {
+         league,
+         draft,
+         watchlist,
+         user,
+         draftPicks,
+         players,
+         team,
+         teams,
+         leagueRules,
+         leagueScoring,
+         draftedPlayers,
+         timerDuration,
+      };
+   }
+);
 
 const Draft = async ({
    params,
@@ -26,46 +106,24 @@ const Draft = async ({
    params: { id: string; draftId: string };
 }) => {
    const supabase = createClient();
+   const {
+      league,
+      draft,
+      watchlist,
+      user,
+      draftPicks,
+      players,
+      team,
+      teams,
+      leagueRules,
+      leagueScoring,
+      draftedPlayers,
+      timerDuration,
+   } = await getDraftProps({ draftId: params.draftId, id: params.id });
 
-   const draft: Awaited<Draft> = await fetchDraftById(supabase, params.draftId);
-   const user: Awaited<User | null> = await getUser(supabase);
-   if (!user) return;
-
-   const watchlist: Awaited<Watchlist> = await fetchWatchlist(
-      supabase,
-      user.id,
-      draft
-   );
-   const draftPicks: Awaited<DraftPick[]> = await fetchDraftPicks(
-      supabase,
-      params.draftId
-   );
-   const league: Awaited<League> = await fetchLeague(supabase, params.id);
-   const players: Awaited<Player[]> = await getPlayers(params.id);
-   const team: Awaited<Team> = await fetchTeam(supabase, user.id, params.id);
-   const leagueRules: Awaited<LeagueRules> = await fetchLeagueRules(
-      supabase,
-      league
-   );
-
-   const leagueScoring: Awaited<LeagueScoring> = await fetchLeagueScoring(
-      supabase,
-      league
-   );
-
-   const teams: Awaited<Team[]> = await fetchTeams(
-      supabase,
-      league.league_id as string
-   );
-
-   const draftedPlayers: Awaited<DraftSelection[]> = await fetchDraftedPlayers(
-      supabase,
-      draft
-   );
-   const timerDuration: Awaited<number> = await getTimerDuration(
-      supabase,
-      leagueRules.id
-   );
+   if (!players.length) {
+      revalidatePath(`/leagues/${params.id}/draft/${params.draftId}`);
+   }
 
    const boardProps: BoardProps = {
       league: league,
